+++ packages/components/src/components.d.ts
@@ -458,7 +458,7 @@ export namespace Components {
          */
         "target"?: LinkTarget;
         /**
-          * Adapts the icon color when used on dark background.
+          * Adapts the link color when used on dark background.
          */
         "theme"?: Theme;
     }
@@ -1541,7 +1541,7 @@ declare namespace LocalJSX {
          */
         "target"?: LinkTarget;
         /**
-          * Adapts the icon color when used on dark background.
+          * Adapts the link color when used on dark background.
          */
         "theme"?: Theme;
     }