+++ packages/components/CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Unreleased]
 
 #### Changed
+- Improved `:focus-visible` and `:hover:focus-visible` colors for `Link Social` and `Link`
+- Improved slotted `<a>` coloring in dark theme for `Link Social` and `Link`
 - Validation messages of `Checkbox Wrapper`, `Radio Button Wrapper`, `Select Wrapper`, `Textarea Wrapper` and `Text Field Wrapper` have now an additional icon representing the validation state.
+
+#### Fixed
+- Slotted `<a>` coloring in dark theme for `Text`, `Headline`, `Text List`, `Banner`, `Select Wrapper` and `Link Pure`
 
 ### [2.3.0] - 2021-07-28